# dynamic-form-builder-engine/docs/swagger.yaml
definitions:
  auth.authResponse:
    properties:
      token:
        type: string
    type: object
  auth.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  formengine.FieldError:
    properties:
      field:
        description: e.g. "/title"; "(root)" for object-level errors like missing-required
        type: string
      message:
        type: string
    type: object
  forms.FormResponse:
    properties:
      createdAt:
        type: string
      currentVersionId:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
  forms.VersionResponse:
    properties:
      changeSummary:
        type: string
      createdAt:
        type: string
      formId:
        type: string
      id:
        type: string
      schemaJson:
        type: object
    type: object
  forms.createFormRequest:
    properties:
      description:
        type: string
      name:
        type: string
      schema:
        type: object
    required:
    - name
    - schema
    type: object
  forms.createVersionRequest:
    properties:
      changeSummary:
        type: string
      schemaJson:
        type: object
    required:
    - schemaJson
    type: object
  response.ErrorResponse:
    properties:
      code:
        type: integer
      errors:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
  response.Page-forms_FormResponse:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      items:
        items:
          $ref: '#/definitions/forms.FormResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Page-submissions_SubmissionResponse:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      items:
        items:
          $ref: '#/definitions/submissions.SubmissionResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Page-users_userResponse:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      items:
        items:
          $ref: '#/definitions/users.userResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  submissions.PublicFormResponse:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      schemaJson:
        type: object
      versionId:
        type: string
    type: object
  submissions.SubmissionResponse:
    properties:
      createdAt:
        type: string
      data:
        type: object
      formId:
        type: string
      formVersionId:
        type: string
      id:
        type: string
      status:
        type: string
      submittedBy:
        type: string
    type: object
  submissions.submitRequest:
    properties:
      data:
        type: object
    required:
    - data
    type: object
  submissions.validationErrorResponse:
    properties:
      code:
        type: integer
      errors:
        items:
          $ref: '#/definitions/formengine.FieldError'
        type: array
      message:
        type: string
    type: object
  users.createUserRequest:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  users.userResponse:
    properties:
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 'Configuration-driven form engine: define, validate, and store dynamic
    forms.'
  title: Dynamic Form Builder API
  version: "1.0"
paths:
  /api/v1/auth/sign-in:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.authResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Log in and receive a JWT
      tags:
      - auth
  /api/v1/forms:
    get:
      parameters:
      - default: 1
        description: Page number (1-based)
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Page-forms_FormResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List the authenticated user's forms
      tags:
      - forms
    post:
      consumes:
      - application/json
      parameters:
      - description: Form + initial schema
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/forms.createFormRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/forms.FormResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a form with its first version
      tags:
      - forms
  /api/v1/forms/{id}:
    get:
      parameters:
      - description: Form ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forms.FormResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a form by ID
      tags:
      - forms
  /api/v1/forms/{id}/submissions:
    get:
      parameters:
      - description: Form ID
        in: path
        name: id
        required: true
        type: string
      - description: Filter by a specific version
        in: query
        name: version_id
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Page-submissions_SubmissionResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List responses for a form (owner only)
      tags:
      - submissions
  /api/v1/forms/{id}/submissions/{submissionId}:
    get:
      parameters:
      - description: Form ID
        in: path
        name: id
        required: true
        type: string
      - description: Submission ID
        in: path
        name: submissionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/submissions.SubmissionResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a single response (owner only)
      tags:
      - submissions
  /api/v1/forms/{id}/versions:
    post:
      consumes:
      - application/json
      parameters:
      - description: Form ID
        in: path
        name: id
        required: true
        type: string
      - description: New schema
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/forms.createVersionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/forms.VersionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new version of a form (becomes current)
      tags:
      - forms
  /api/v1/forms/{id}/versions/current:
    get:
      parameters:
      - description: Form ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forms.VersionResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get the current (live) version of a form
      tags:
      - forms
  /api/v1/public/forms/{id}:
    get:
      parameters:
      - description: Form ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/submissions.PublicFormResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get a form's current schema for public rendering
      tags:
      - public
  /api/v1/public/forms/{id}/submissions:
    post:
      consumes:
      - application/json
      description: Validates the payload against the form's CURRENT version and pins
        the submission to it.
      parameters:
      - description: Form ID
        in: path
        name: id
        required: true
        type: string
      - description: Submission data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/submissions.submitRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/submissions.SubmissionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/submissions.validationErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Submit a response to a form (anonymous, public)
      tags:
      - public
  /api/v1/users:
    get:
      parameters:
      - default: 1
        description: Page number (1-based)
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Page-users_userResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: List all users
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: User registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.createUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/users.userResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Register a new user
      tags:
      - users
  /api/v1/users/{id}:
    get:
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.userResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get a user by ID
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your JWT.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"